lua-mod: add more of the amixer API to lua-mod
This patch adds additional amixer API calls to lua-mod.
Specifically, this is to expose enum controls, to set and also
to view a list of enum names. This is so that we can change
various simple mixer controls provided by ST, which are used to
control the DSP on Lawrence.

Signed-off-by: Ben Collins <ben.collins@garmin.com>